requirements.txt
# FastAPI Web框架
fastapi>=0.116.1

# ASGI服务器
uvicorn>=0.35.0

# 浏览器自动化
playwright>=1.55.0

# 异步ORM框架
tortoise-orm>=0.25.1
aiosqlite>=0.19.0  # SQLite异步驱动

# 环境变量管理
python-dotenv>=0.19.0

# 开发依赖
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0

# HTTP客户端(可选，用于测试)
requests>=2.31.0
